+++ src/Data/Deriving.hs
@@ -79,25 +79,25 @@ The following changes have been backported:
 * In GHC 8.4, deriving 'Functor' and 'Traverable' was changed so that it uses 'coerce'
   for efficiency when the last parameter of the data type is at phantom role.
 
-* In GHC 8.4, the `EmptyDataDeriving` proposal brought forth a slew of changes related
+* In GHC 8.4, the @EmptyDataDeriving@ proposal brought forth a slew of changes related
   to how instances for empty data types (i.e., no constructors) were derived. These
   changes include:
 
-    * For derived `Eq` and `Ord` instances for empty data types, simply return
-      `True` and `EQ`, respectively, without inspecting the arguments.
+    * For derived 'Eq' and 'Ord' instances for empty data types, simply return
+      'True' and 'EQ', respectively, without inspecting the arguments.
 
-    * For derived `Read` instances for empty data types, simply return `pfail`
-      (without `parens`).
+    * For derived 'Read' instances for empty data types, simply return 'pfail'
+      (without 'parens').
 
-    * For derived `Show` instances for empty data types, inspect the argument
-      (instead of `error`ing).
+    * For derived 'Show' instances for empty data types, inspect the argument
+      (instead of 'error'ing).
 
-    * For derived `Functor` and `Traversable` instances for empty data
-      types, make `fmap` and `traverse` strict in its argument.
+    * For derived 'Functor' and 'Traversable' instances for empty data
+      types, make 'fmap' and 'traverse' strict in its argument.
 
-    * For derived `Foldable` instances, do not error on empty data types.
-      Instead, simply return the folded state (for `foldr`) or `mempty` (for
-      `foldMap`), without inspecting the arguments.
+    * For derived 'Foldable' instances, do not error on empty data types.
+      Instead, simply return the folded state (for 'foldr') or 'mempty' (for
+      'foldMap'), without inspecting the arguments.
 
 * In GHC 8.6, the @DerivingVia@ language extension was introduced.
   @deriving-compat@ provides an interface which attempts to mimic this
@@ -107,8 +107,11 @@ The following changes have been backported:
   Since the generated code requires the use of @TypeApplications@, this can
   only be backported back to GHC 8.2.
 
-* In GHC 8.6, deriving `Read` was changed so as to factor out certain commonly
+* In GHC 8.6, deriving 'Read' was changed so as to factor out certain commonly
   used subexpressions, which significantly improve compliation times.
+
+* In GHC 8.10, @DerivingVia@ permits \"floating\" type variables in @via@ types,
+  such as the @a@ in @'deriveVia' [t| forall a. Show MyInt ``Via`` Const Int a |]@.
 -}
 
 {- $derive