+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### next [????.??.??]
+* Permit `deriveVia` to use "floating" `via` type variables, such as the `a`
+  in:
+
+  ```hs
+  deriveVia [t| forall a. Show MyInt `Via` Const Int a |]
+  ```
+
 ### 0.5.6 [2019.05.02]
 * Support deriving `Eq`, `Ord`, and `Show` instances for data types with fields
   of type `Int8#`, `Int16#`, `Word8#`, or `Word16#` on GHC 8.8 or later.